Fix compilation on MacOS.
The error previously was (on one line)
```
dyld: BIND_OPCODE_SET_SEGMENT_AND_OFFSET_ULEB has segment 1
which is not writable in /<redacted>/examples/./a.out
```

The issue was that Cranelift is configured by default not to use PIC
(position-independent code). This causes problems because MacOS
does not seem to like the use of absolute addresses. Changing this
configuration resolved the bug.